Event

+++ LDA/include/LdaBin.h
@@ -18,7 +18,7 @@
 
 struct CSTaintBin
 {
-	unsigned InstID;
+	unsigned long InstID;
 	unsigned InTaintBits;
 	unsigned OutTaintBits;
 	unsigned CalleeNum;
@@ -31,8 +31,8 @@ struct FldaBin
     unsigned FuncId;
 	unsigned TaintInstNum;
 	unsigned TaintCINum;
-	//unsigned InstID[]
-	//unsigned TaintCI[] 
+	//unsigned long InstID[]
+	//CSTaintBin TaintCI[] 
 };
 
 struct LdaBin
